+++ src/main/java/com/example/sns/like/domain/LikeRepository.java
@@ -1,5 +1,6 @@
-package com.example.sns.post.domain;
+package com.example.sns.like.domain;
 
+import com.example.sns.like.domain.Like;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
