+++ src/main/java/com/example/sns/post/application/FeedService.java
@@ -1,6 +1,6 @@
 package com.example.sns.post.application;
 
-import com.example.sns.post.domain.LikeRepository;
+import com.example.sns.like.domain.LikeRepository;
 import com.example.sns.post.domain.Post;
 import com.example.sns.post.domain.PostRepository;
 import com.example.sns.post.presentiation.dto.MyFeedResponse;
